+++ src/main/java/de/maxhenkel/plane/entity/EntityBushPlane.java
@@ -7,13 +7,14 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.phys.Vec3;
+import org.jetbrains.annotations.Nullable;
 
 public class EntityBushPlane extends EntityPlaneBase {
 
     public EntityBushPlane(Level world) {
         this(Main.BUSH_PLANE_ENTITY_TYPE.get(), world);
-        lootTable = ResourceKey.create(Registries.LOOT_TABLE, ResourceLocation.fromNamespaceAndPath(Main.MODID, "entities/bush_plane_" + getPlaneType().getTypeName()));
     }
 
     public EntityBushPlane(EntityType<?> type, Level world) {
@@ -48,6 +49,11 @@ public class EntityBushPlane extends EntityPlaneBase {
     @Override
     public Vec3[] getPlayerOffsets() {
         return new Vec3[]{new Vec3(0D, 0D, 0.5D)};
+    }
+
+    @Override
+    public ResourceKey<LootTable> getLootTable() {
+        return ResourceKey.create(Registries.LOOT_TABLE, ResourceLocation.fromNamespaceAndPath(Main.MODID, "entities/bush_plane_" + getPlaneType().getTypeName()));
     }
 
 }