+++ src/main/java/de/maxhenkel/plane/entity/EntityCargoPlane.java
@@ -21,6 +21,7 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.ChestMenu;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.phys.Vec3;
 import net.neoforged.neoforge.network.PacketDistributor;
 
@@ -37,7 +38,6 @@ public class EntityCargoPlane extends EntityPlaneBase {
     public EntityCargoPlane(EntityType<?> type, Level world) {
         super(type, world);
         cargoInventory = new SimpleContainer(54);
-        lootTable = ResourceKey.create(Registries.LOOT_TABLE, ResourceLocation.fromNamespaceAndPath(Main.MODID, "entities/cargo_plane_" + getPlaneType().getTypeName()));
     }
 
     @Override
@@ -119,6 +119,11 @@ public class EntityCargoPlane extends EntityPlaneBase {
     @Override
     public Vec3[] getPlayerOffsets() {
         return new Vec3[]{new Vec3(0.5D, 0D, 0.8D), new Vec3(-0.5D, 0D, 0.8D)};
+    }
+
+    @Override
+    public ResourceKey<LootTable> getLootTable() {
+        return ResourceKey.create(Registries.LOOT_TABLE, ResourceLocation.fromNamespaceAndPath(Main.MODID, "entities/cargo_plane_" + getPlaneType().getTypeName()));
     }
 
 }