+++ src/main/java/de/maxhenkel/plane/entity/EntityPlane.java
@@ -14,6 +14,7 @@ import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.phys.Vec3;
 import net.neoforged.neoforge.network.PacketDistributor;
 
@@ -27,7 +28,6 @@ public class EntityPlane extends EntityPlaneBase {
 
     public EntityPlane(EntityType<?> type, Level world) {
         super(type, world);
-        lootTable = ResourceKey.create(Registries.LOOT_TABLE, ResourceLocation.fromNamespaceAndPath(Main.MODID, "entities/plane_" + getPlaneType().getTypeName()));
     }
 
     @Override
@@ -75,6 +75,11 @@ public class EntityPlane extends EntityPlaneBase {
     @Override
     public Vec3[] getPlayerOffsets() {
         return new Vec3[]{new Vec3(0D, 0D, 1D), new Vec3(0D, 0D, 0.5D), new Vec3(0D, 0D, 0D)};
+    }
+
+    @Override
+    public ResourceKey<LootTable> getLootTable() {
+        return ResourceKey.create(Registries.LOOT_TABLE, ResourceLocation.fromNamespaceAndPath(Main.MODID, "entities/plane_" + getPlaneType().getTypeName()));
     }
 
 }